fix: Type verifier service
